Change package layout to make ros package release easier (#44)
* move ros/cmake and package xml out. merge core cmake into root cmake

this is basically the lesser evil of the two options i thought up to
make the package build on the ros build farm. the other was to patch the
release repository to manually copy the cpp source files (scripted) as a
patch onto upstream. horrible either way.

* build ros by default, but disable it for the python build. if someone
does a cpp build then they'll have to manually disable it

* the makefile recipe now disables ros build by default for cpp only
build. at least

* clean up ros workflows a bit